raspi/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
# Donanım girişleri için isteğe bağlı; gerçek Pi üzerinde kurulabilir.
RPi.GPIO>=0.7.0
gpiozero>=1.6.2
